agent_results/end2end_gemini/iclr2025_wsl/claude_code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.65.0
einops>=0.6.0
transformers>=4.30.0
tensorboard>=2.12.0
pyyaml>=6.0.0
wandb>=0.15.0
onnx>=1.14.0